Refactor: Remove old test scripts and add new test implementations
- Deleted obsolete test scripts: test_dropdown_visibility.py, test_enable_flow.py, test_refresh_interval_minimum.py, test_spinbox_minimum.py, test_sync_validation.py, test_timing_sync.py.
- Added new test scripts in the tests directory with updated implementations for dropdown visibility, enabling sync and light control, refresh interval enforcement, spinbox minimum value, and synchronization validation.
- Ensured all new tests are structured to validate functionality and UI behavior in the application.
