+++ tests/test_refresh_interval_minimum.py
