+++ tests/test_spinbox_minimum.py
